逻辑相关的先提交一波

+++ opencvIdentification/src/main/java/com/nwq/opencv/db/entity/FunctionEntity.kt
@@ -18,7 +18,7 @@ import com.nwq.opencv.db.IdentifyDatabase
 data class FunctionEntity(
 
     @PrimaryKey(autoGenerate = true)
-    var id: Int = 0,
+    var id: Long = 0,
 
     //识别标签 比如"主菜单","位置菜单"
     var keyTag: String,
